refactor: 장소, 구역과 좌석 컨트롤러 분리
