+++ src/main/java/com/culture/ticketing/place/api/PlaceController.java
@@ -1,12 +1,8 @@
 package com.culture.ticketing.place.api;
 
-import com.culture.ticketing.place.application.AreaService;
 import com.culture.ticketing.place.application.PlaceService;
-import com.culture.ticketing.place.application.SeatService;
-import com.culture.ticketing.place.application.dto.PlaceAreaSaveRequest;
 import com.culture.ticketing.place.application.dto.PlaceResponse;
 import com.culture.ticketing.place.application.dto.PlaceSaveRequest;
-import com.culture.ticketing.place.application.dto.PlaceSeatSaveRequest;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -17,13 +13,9 @@ import java.util.List;
 public class PlaceController {
 
     private final PlaceService placeService;
-    private final SeatService seatService;
-    private final AreaService areaService;
 
-    public PlaceController(PlaceService placeService, SeatService seatService, AreaService areaService) {
+    public PlaceController(PlaceService placeService) {
         this.placeService = placeService;
-        this.seatService = seatService;
-        this.areaService = areaService;
     }
 
     @PostMapping("")
@@ -39,14 +31,4 @@ public class PlaceController {
         return placeService.findPlaces(offset, size);
     }
 
-    @PostMapping("/areas")
-    public void postPlaceArea(@Valid @RequestBody PlaceAreaSaveRequest request) {
-
-        areaService.createPlaceArea(request);
-    }
-    @PostMapping("/seats")
-    public void postPlaceSeat(@Valid @RequestBody PlaceSeatSaveRequest request) {
-
-        seatService.createPlaceSeat(request);
-    }
 }
